*: fix a few sparse warnings
I had to pull sparse from the git repo and apply the __leaf__
attribute patch on top. Also, you need to provide the GCCDIR
if it's different for you.

I just remembered that this tool set is written for 64bit. I
don't think I'll put in the time to support anything else. ;)