sync: refactor alloc and prealloc, plus type style fixes
It was about time to clean up the macro mess in the alloc wrapper
code. The new version depends even more on macros, but the code
should be a lot easier to understand now.

Prealloc lost some weight, because fields count and mem_size were
never being used.

And, finally, type's __CHECKER__ macros, which are only in place for
a complaining sparse(1), have been made easier on the eyes.