regress: regression test infrastructure
As always, thanks go out to the BSD guys for working this out.
This gets rid of the "regress" target in favor of a regress/
directory. Invoking make inside this directory will cause all
available regression tests to run. Erros, if any, will be shown.
If your binary supports different input streams, you can add
different regression test targets and specify their individual
input and output in the .in and .out files, respectively.