load: thin pcap/erf loader
This pulls in the first library element: a trace file loader to avoid
pulling in pcap(3) in its full glory. It really doesn't do much other
than providing a simple interface to pull packet after packet from
the input stream. Unit test is provided. Be sure to take a peek at
the man page as well.