li: merge lightweight inspection
Things are still in flux, but it's time to live up to the
promises!  This merges the current state of our lightweight
inspection code.  We will continue to update protocols as
we encounter them internally.  The peak_li(3) man page is
quite extensive and will help you with integration or how
to add new protocols.  Got any problems or feedback?  Let
us know!