make: s/makefile.inc/common.mk/g
Some internal debates concluded that BSD make runs on most systems
anyway and that we don't really need the GNU make glue.  For now
it stays, but to give way to some hiccups with *.inc file handling
with BSD make, we decided to rename the common parts of the makefiles.