netmap: bring in the latest API adoption
Luigi discontinued the original nm_util files so we imported the
old code into our library and adapted it to work with the current
API version 11.  Somewhat tested, but there may be final quirks
to sort out.  This is *not* compatible with the FreeBSD 10 release,
but maybe the API has made its way to stable.  I don't know.

Kudos to Alexey for working through this!  :)