netmap: improve transparent mode responsiveness
Also poll for TX, which flushes said queues in transparent mode.
It's a bit weird that RX poll is in charge of transparent mode,
handing packets over to TX without issuing an immediate flush.

It's easier to fix this in userland with a one liner, but ultimately
the netmap(4) code needs to be improved.  I don't think the
transparent mode is widely used (if at all), so it's only natural
that this is the current state.

Also set timestamping for every ring as there are going to be
a couple of improvements on that front as well.  For now this is
a nop unless the dev.netmap.no_timestamp sysctl is set to `1'.